+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Changes that have landed in develop but are not yet released.
 
+## v1.3.0 (Feb 20, 2020)
+- Fix withTooltip so that choose language will close withTooltip
+- Quick hack to full reload the story (support the use of global i18next)
+
 ## v1.1.0 (Mar 8, 2019)
 
 ### Breaking Changes